chore(refactor): refactors go code
