+++ internal/analytics/types.go
